Add new options to the checksum checker drush cmd
The checksum checker options ask the user to specify the number of items
per invocation of the batch process. This is difficult to estimate if
your repository is continually changing in size.

This pull request adds extra optional arguements to the checksum checker
drush command that let the checksum checker determine the limit based on
the number of items in your repository and how long you want it to take
to check all your checksums.