Queuing improvements (#50)
* Ensure 'skip on cron' is set properly in hook_cron_queue_info

When the 'cron method' configuration is not set to 'drupal',
ensure we don't process the queue with cron. With this configuration
the queue shouldn't get populated when cron runs anyway, but if the
queue did somehow contain items we would run into issues.

* Explicitly return a reliable queue

DrupalQueue::get() does return a SystemQueue by default, but this
guards against other modules/configurations potentially changing
that behavior.

* Ensure that drupal caches a reliable queue

DrupalQueue::get() uses a static variable to store queues,
so the first call to that function should pass reliable=TRUE
in order to ensure we get a reliable queue.